model 세분화
