+++ application/models/Category.py
@@ -1,0 +1,12 @@
+from application import db
+
+
+class Category(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(50), nullable=False)
+
+    def __repr__(self):
+        return '<Category name : %r>' % self.name
+
+    def get_name(self):
+        return self.name
