+++ application/models/Post.py
@@ -1,4 +1,3 @@
-from flask_sqlalchemy import SQLAlchemy
 from application import db
 import datetime
 
@@ -18,14 +17,3 @@ class Post(db.Model):
 
     def get_title(self):
         return self.title
-
-
-class Category(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(50), nullable=False)
-
-    def __repr__(self):
-        return '<Category name : %r>' % self.name
-
-    def get_name(self):
-        return self.name
